Remove use of SinglePathCatalog
replaced with use of inner class that only functions to allow the
catalog to be used in try-with-resources statements.
